requirements.txt
asn1crypto==0.22.0
attrs==21.4.0
Automat==20.2.0
certifi==2017.4.17
cffi==1.15.0
chardet==3.0.4
constantly==15.1.0
cryptography==36.0.1
cssselect==1.0.1
Django==1.11.17
django-filter==1.0.4
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.6.3
djangorestframework-filters==0.10.1
enum34==1.1.6
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==17.2.1
idna==2.5
incremental==21.3.0
ipaddress==1.0.18
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.7.1
mongoengine==0.23.1
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
pymongo==3.12.3
pyOpenSSL==21.0.0
pytz==2017.2
queuelib==1.4.2
requests==2.18.2
Scrapy==2.5.1
service-identity==17.0.0
six==1.10.0
Twisted==21.7.0
typing_extensions==4.0.1
urllib3==1.22
w3lib==1.22.0
zope.interface==5.4.0
